defautl hostname url set-up
