+++ .github/workflows/ios-distribution.yml
@@ -16,7 +16,7 @@ jobs:
         with:
           ssh-key: ${{ secrets.ACTION_DEPLOY_KEY }}
       # We need Xcode to build the iOS app
-      - uses: maxim-lobanov/setup-xcode@v1
+      - uses: maxim-lobanov/setup-xcode@v1.6.0
         with:
           xcode-version: 15.3
       - name: set up JDK 17
@@ -64,6 +64,7 @@ jobs:
           cd iosApp
 
           xcrun xcodebuild \
+            -workspace "iosApp.xcworkspace" \
             -scheme "iosApp" \
             -configuration "Release" \
             -sdk "iphoneos" \
